tests/requirements.txt
pytest
datadog-checks-base[deps]
requests
